+++ src/pages/todo/Todos.js
@@ -1,19 +1,13 @@
 import React, { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useState } from 'react';
-import axios from 'axios';
 import styled from 'styled-components';
 import Todo from './Todo';
-import { API } from '../../config';
+import { TodoAxios } from '../../api/TodoAxios';
 
 const Todos = () => {
   const [todoInput, setTodoInput] = useState('');
   const [todos, setTodos] = useState([]);
-
-  const access_token = localStorage.getItem('token');
-  const config = {
-    headers: { Authorization: 'Bearer ' + access_token },
-  };
 
   const navigate = useNavigate();
 
@@ -21,15 +15,7 @@ const Todos = () => {
 
   const plusTodo = async e => {
     e.preventDefault();
-    const res = await axios.post(
-      `${API.Todo}`,
-      {
-        todo: todoInput,
-        isCompleted: false,
-      },
-      config
-    );
-    const { data } = res;
+    const data = await TodoAxios.ADD(todoInput);
     setTodos(pre => [...pre, data]);
     setTodoInput('');
   };
@@ -60,48 +46,28 @@ const Todos = () => {
   const TodoModify = async (e, checkedId, setToogle) => {
     e.preventDefault();
     const [targetValue] = todos.filter(({ id }) => id === Number(checkedId));
-    await axios.put(
-      `${API.Todo}/${checkedId}`,
-      {
-        todo: targetValue.todo,
-        isCompleted: targetValue.isCompleted,
-      },
-      config
-    );
+    await TodoAxios.PUT(checkedId, targetValue.todo, targetValue.isCompleted);
     setToogle(pre => !pre);
   };
 
   const TodoDelete = async (e, checkedId, setToogle) => {
     e.preventDefault();
-    await axios.delete(`${API.Todo}/${checkedId}`, config);
+    await TodoAxios.DELETE(checkedId);
+    setTodos(prev => prev.filter(({ id }) => id !== Number(checkedId)));
     setToogle(pre => !pre);
     alert('삭제 완료');
-    window.location.reload(true);
   };
 
-  const TodoChange = e => {
+  // TODO : 취소 버튼부터 리팩토링!
+
+  const TodoChange = (e, setToogle) => {
     e.preventDefault();
-    alert('취소되었습니다');
-    window.location.reload(true);
+    setToogle(pre => !pre);
   };
 
   useEffect(() => {
-    const getTodoList = async () => {
-      if (!access_token) {
-        alert('로그인이 필요합니다');
-        navigate('/');
-      } else {
-        try {
-          const res = await axios.get(`${API.Todo}`, config);
-          const { data } = res;
-          setTodos(data);
-        } catch (err) {
-          throw new Error(err);
-        }
-      }
-    };
-    getTodoList();
-  }, []);
+    TodoAxios.GET(navigate, setTodos);
+  }, [navigate]);
 
   return (
     <Layout>
